refactor: move function related to paritions into separate file
this makes it a bit easier to navigate the start.go file by moving
some of the disk size and partition modification functions into  a
separate disks.go file in the same machine package
